requirements.txt
python-dotenv>=1.0.0
faiss-cpu==1.7.3
torchcrepe==0.0.20
praat-parselmouth>=0.4.2
ffmpeg-python>=0.2.0
pyworld==0.3.2
fairseq==0.12.2
pip==22.0
numpy==1.23.5
av